I worked on Trainings app: I reorganized the tests in different files (1 for trainings, 1 for sections, 1 for chapters
